# Closes #1300 - Improve Performance of DataFrame Display