--- a/ServerModules.cfg
+++ b/ServerModules.cfg
@@ -5,6 +5,7 @@ ArraySetopsMsg
 KExtremeMsg
 ArgSortMsg
 SegmentedMsg
+DataFrameIndexingMsg
 OperatorMsg
 RandMsg
 IndexingMsg