--- a/src/DataFrameIndexingMsg.chpl
+++ b/src/DataFrameIndexingMsg.chpl
@@ -8,6 +8,7 @@ module DataFrameIndexingMsg
     use Logging;
     use Message;
 	use SegmentedMsg;
+	use AryUtil;
 
     use MultiTypeSymEntry;
     use MultiTypeSymbolTable;
@@ -18,7 +19,7 @@ module DataFrameIndexingMsg
     const dfiLogger = new Logger(logLevel);
 
 	// gather indexing by integer index vector
-	proc dfIdxHelper(idx: borrowed SymEntry(int), columnVals: borrowed SymEntry(?t), st: borrowed SymTab, rtnName: bool=false): string throws {
+	proc dfIdxHelper(idx: borrowed SymEntry(int), columnVals: borrowed SymEntry(?t), st: borrowed SymTab, col: string, rtnName: bool=false): string throws {
 		param pn = Reflection.getRoutineName();
 		// get next symbol name
         var rname = st.nextName();
@@ -52,8 +53,39 @@ module DataFrameIndexingMsg
 			return rname;
 		}
 
-		var repMsg =  "pdarray+created " + st.attrib(rname);
+		var repMsg =  "pdarray+%s+created %s".format(col, st.attrib(rname));
 		return repMsg;
+	}
+
+	proc df_seg_array_idx(idx: borrowed SymEntry(int), segments: borrowed SymEntry(int), values: borrowed SymEntry(?t), col: string, st: borrowed SymTab): string throws {
+		var lens: [0..#idx.size] int;
+		var orig_segs: [0..#idx.size] int = segments.a[idx.a];
+
+		const ref high = orig_segs.domain.high;
+		forall (i, os, l) in zip(orig_segs.domain, orig_segs, lens){
+			if(i == high) {
+				l = values.size - os;
+			} else {
+				l = orig_segs[i+1] - os;
+			}
+		}
+
+		var rvals: [0..#(+ reduce lens)] t;
+		var rsegs = (+ scan lens) - lens;
+
+		forall(i, rs, os, l) in zip(orig_segs.domain, rsegs, orig_segs, lens){
+			var v = new lowLevelLocalizingSlice(values.a, os..#l);
+			for j in 0..#l{
+				rvals[rs+j] = v.ptr[j];
+			}
+		}
+
+		var s_name = st.nextName();
+		st.addEntry(s_name, new shared SymEntry(rsegs));
+		var v_name = st.nextName();
+		st.addEntry(v_name, new shared SymEntry(rvals));
+
+		return "SegArray+%s+created %s+created %s".format(col, s_name, v_name);
 	}
 
     proc dataframeBatchIndexingMsg(cmd: string, payload: string, st: borrowed SymTab): MsgTuple throws {
@@ -81,16 +113,17 @@ module DataFrameIndexingMsg
 
 		forall (i, rpm, ele) in zip(repMsgList.domain, repMsgList, eleList) { 
 			var ele_parts = ele.split("+");
+			ref col_name = ele_parts[1];
 			select (ele_parts[0]) {
 				when ("Categorical") {
 					dfiLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),"Element at %i is Categorical".format(i));
-					ref codes_name = ele_parts[1];
-					ref categories_name = ele_parts[2];
+					ref codes_name = ele_parts[2];
+					ref categories_name = ele_parts[3];
 					dfiLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),"Codes Name: %s, Categories Name: %s".format(codes_name, categories_name));
 
 					var gCode: borrowed GenSymEntry = getGenericTypedArrayEntry(codes_name, st);
 					var code_vals = toSymEntry(gCode, int);
-					var idxCodeName = dfIdxHelper(idx, code_vals, st, true);
+					var idxCodeName = dfIdxHelper(idx, code_vals, st, col_name, true);
 					
 					var args: [1..2] string = [categories_name, idxCodeName];
 					var repTup = segPdarrayIndex("str", args, st);
@@ -98,37 +131,37 @@ module DataFrameIndexingMsg
 						throw new IllegalArgumentError(repTup.msg);
 					}
 
-					rpm = "Strings+%s,".format(repTup.msg);
+					rpm = "%jt".format("Strings+%s+%s".format(col_name, repTup.msg));
 				}
 				when ("Strings") {
 					dfiLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),"Element at %i is Strings".format(i));
-					var args: [1..2] string = [ele_parts[1], iname];
+					var args: [1..2] string = [ele_parts[2], iname];
 					var repTup = segPdarrayIndex("str", args, st);
 					if repTup.msgType == MsgType.ERROR {
 						throw new IllegalArgumentError(repTup.msg);
 					}
 
-					rpm = "Strings+%s,".format(repTup.msg);
+					rpm = "%jt".format("Strings+%s+%s".format(col_name, repTup.msg));
 				}
 				when ("pdarray"){
 					dfiLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),"Element at %i is pdarray".format(i));
-					var gCol: borrowed GenSymEntry = getGenericTypedArrayEntry(ele_parts[1], st);
+					var gCol: borrowed GenSymEntry = getGenericTypedArrayEntry(ele_parts[2], st);
 					select (gCol.dtype) {
 						when (DType.Int64) {
 							var col_vals = toSymEntry(gCol, int);
-							rpm = "%s,".format(dfIdxHelper(idx, col_vals, st));
+							rpm = "%jt".format(dfIdxHelper(idx, col_vals, st, col_name));
 						}
 						when (DType.UInt64) {
 							var col_vals = toSymEntry(gCol, uint);
-							rpm = "%s,".format(dfIdxHelper(idx, col_vals, st));
+							rpm = "%jt".format(dfIdxHelper(idx, col_vals, st, col_name));
 						}
 						when (DType.Bool) {
 							var col_vals = toSymEntry(gCol, bool);
-							rpm = "%s,".format(dfIdxHelper(idx, col_vals, st));
+							rpm = "%jt".format(dfIdxHelper(idx, col_vals, st, col_name));
 						}
 						when (DType.Float64){
 							var col_vals = toSymEntry(gCol, real);
-							rpm = "%s,".format(dfIdxHelper(idx, col_vals, st));
+							rpm = "%jt".format(dfIdxHelper(idx, col_vals, st, col_name));
 						}
 						otherwise {
 							var errorMsg = notImplementedError(pn,dtype2str(gCol.dtype));
@@ -138,6 +171,19 @@ module DataFrameIndexingMsg
 					}
 				}
 				// TODO - when SegArray
+				when ("SegArray"){
+					dfiLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),"Element at %i is SegArray".format(i));
+					ref segments_name = ele_parts[2];
+					ref values_name = ele_parts[3];
+					dfiLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),"Segments Name: %s, Values Name: %s".format(segments_name, values_name));
+
+					var gSeg: borrowed GenSymEntry = getGenericTypedArrayEntry(segments_name, st);
+					var segments = toSymEntry(gSeg, int);
+					var gVal: borrowed GenSymEntry = getGenericTypedArrayEntry(values_name, st);
+					var values = toSymEntry(gVal, int);
+
+					rpm = "%jt".format(df_seg_array_idx(idx, segments, values, col_name, st));
+				}
 				otherwise {
 					var errorMsg = notImplementedError(pn, ele_parts[0]);
 					dfiLogger.error(getModuleName(),getRoutineName(),getLineNumber(),errorMsg);