Closes #1300 

- Updates the `_get_head_tail()` method to function by creating a single server message instead of 1 per column in the DataFrame. This was done by adding `_get_head_tail_server()` which allows us to benchmark the new code against the old. The old function has been retained for benchmarking.
- Adds a `DataFrameIndexingMsg.chpl` to parse and process the server message for indexing the columns. The message is configured so the the column type, column name (in the df), column object name (server-side) are sent to the server to indexing to the appropriate head/tail. 
- Benchmarking configured to test `_rep_html()` which now calls `_get_head_tail_server()`. The benchmark also checks `_get_head_tail_server()` and `_get_head_tail()` directly to allow for easy comparison. 

Per the issue description, @joshmarshall1 and I did this without using aggregation because the datasets should be relatively small.

Benchmarking on a single node shows an performance already, but we will need to benchmark this on a multi-node system. Results from single node:

```bash
array size = 10,000
number of trials =  5
>>> arkouda dataframe display
numLocales = 1, N = 10,000
  _repr_html_ Average time = 0.0401 sec
  _repr_html_ Average rate = 0.03 GiB/sec
  _get_head_tail_server Average time = 0.0406 sec
  _get_head_tail_server Average rate = 0.03 GiB/sec
  _get_head_tail Average time = 0.0825 sec
  _get_head_tail Average rate = 0.01 GiB/sec

```

@reuster986 - I will leave it up to you if you would like to request an out of band benchmark.