@Ethan-DeBandi99 2022-04-28
Updated benchmark results on single node with requested updates from @reuster986 implemented.

```bash
array size = 10,000
number of trials =  5
>>> arkouda dataframe display
numLocales = 1, N = 10,000
  _get_head_tail_server Average time = 0.0261 sec
  _get_head_tail_server Average rate = 0.05 GiB/sec
  _get_head_tail Average time = 0.0547 sec
  _get_head_tail Average rate = 0.02 GiB/sec
```