@Ethan-DeBandi99 2022-04-29
@reuster986 not going to lie, I am a bit confused as to why a list can't index into rows. Can you give me some details on your thought process here? I will get it pulled out though